Middleware for verifyTocken
Middleware added for verifyTocken
#33 solved.
